+++ src/Application.php
@@ -12,23 +12,28 @@ namespace Turbine;
 
 use League\Container\ContainerAwareInterface;
 use League\Container\ContainerInterface;
+use League\Container\Exception\NotFoundException;
 use League\Container\ReflectionContainer;
 use League\Event\Emitter;
 use League\Event\EmitterInterface;
 use League\Event\EmitterTrait;
 use League\Event\ListenerAcceptorInterface;
-use League\Route\Strategy\RequestResponseStrategy;
 use Turbine\Psr7\HttpKernelInterface;
 use Turbine\Psr7\TerminableInterface;
-use Turbine\Route\Strategy\WireableStrategy;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use League\Container\Container;
 use League\Route\RouteCollection;
 use Monolog\Logger;
+use Whoops\Handler\HandlerInterface;
+use Whoops\Handler\JsonResponseHandler;
+use Whoops\Handler\PlainTextHandler;
+use Whoops\Handler\PrettyPageHandler;
+use Whoops\Run;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\Response\HtmlResponse;
 use Zend\Diactoros\Response\JsonResponse;
+use Zend\Diactoros\Response\TextResponse;
 use Zend\Diactoros\ServerRequestFactory;
 
 /**
@@ -37,6 +42,9 @@ use Zend\Diactoros\ServerRequestFactory;
 class Application implements ApplicationInterface, ContainerAwareInterface, HttpKernelInterface, ListenerAcceptorInterface, TerminableInterface, \ArrayAccess
 {
     use EmitterTrait;
+
+    const CONFIG_ERROR = true;
+    const CONFIG_DEBUG = true;
 
     /**
      * @var \League\Route\RouteCollection
@@ -70,30 +78,84 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     protected $responseEmitter;
 
     /**
+     * Set while handle exception.
+     * @var bool
+     */
+    private $error = false;
+
+    /**
      * New Application.
      *
-     * @param bool $debug Enable debug mode
+     * @param bool|array $configuration Enable debug mode
      */
-    public function __construct($debug = true)
+    public function __construct($configuration = [])
     {
-        $this->setConfig('debug', $debug);
-
-        $this->setExceptionDecorator(function (\Exception $e) {
-
-            $body = [
-                'error' => [
-                    'message' => $e->getMessage()
-                ]
-            ];
-
-            if ($this->getConfig('debug', true) === true) {
-                $body['error']['trace'] = explode(PHP_EOL, $e->getTraceAsString());
+        if (is_bool($configuration)) {
+            $this->setConfig('debug', $configuration);
+        } elseif (is_array($configuration)) {
+            foreach ($configuration as $key => $value) {
+                $this->setConfig($key, $value);
             }
+        }
 
-            $response = new JsonResponse($body, method_exists($e, 'getStatusCode') ? $e->getStatusCode() : 500);
+        if (!isset($this->config['debug'])) {
+            $this->setConfig('debug', static::CONFIG_DEBUG);
+        }
+        if (!isset($this->config['debug'])) {
+            $this->setConfig('errors', static::CONFIG_ERROR);
+        }
 
-            return $response;
-        });
+    }
+
+    /**
+     * Set a config item. Add recursive if key is traversable.
+     *
+     * @param string|array|\Traversable $key
+     * @param mixed $value
+     *
+     * @return $this
+     */
+    public function setConfig($key, $value = null)
+    {
+        if (is_array($key) || $key instanceof \Traversable) {
+            $config = $key;
+            foreach ($config as $key => $value) {
+                $this->setConfig($key, $value);
+            }
+        } else {
+            $this->config[$key] = $value;
+        }
+
+        return $this;
+    }
+
+    /**
+     * Get a config key's value
+     *
+     * @param string $key
+     * @param mixed $default
+     *
+     * @return mixed
+     */
+    public function getConfig($key = null, $default = null)
+    {
+        if ($key === null) {
+            return $this->config;
+        }
+
+        return $this->hasConfig($key) ? $this->config[$key] : $default;
+    }
+
+    /**
+     * Check if key exists
+     *
+     * @param $key
+     *
+     * @return bool
+     */
+    public function hasConfig($key)
+    {
+        return isset($this->config[$key]);
     }
 
     /**
@@ -121,7 +183,7 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     /**
      * Get the container.
      *
-     * @return \League\Container\ContainerInterface
+     * @return \League\Container\Container
      */
     public function getContainer()
     {
@@ -133,31 +195,38 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     }
 
     /**
-     * Return the router.
-     *
-     * @return \League\Route\RouteCollection
-     */
-    public function getRouter()
-    {
-        if (!isset($this->router)) {
-            $this->router = (new RouteCollection($this->getContainer()));
-        }
-
-        return $this->router;
-    }
-
-    /**
      * Get the Emitter.
      *
      * @return EmitterInterface
      */
     public function getEmitter()
     {
-        if (! $this->emitter) {
+        if (!$this->emitter) {
             $this->emitter = new Emitter();
         }
 
         return $this->emitter;
+    }
+
+    /**
+     * Get the response
+     *
+     * @return HandlerInterface
+     */
+    public function getErrorResponseHandler()
+    {
+        if (!$this->getContainer()->has(HandlerInterface::class)) {
+            if ($this->isCli()) {
+                $class = PlainTextHandler::class;
+            } elseif ($this->isAjax()) {
+                $class = JsonResponseHandler::class;
+            } else {
+                $class = PrettyPageHandler::class;
+            }
+            $this->getContainer()->add(HandlerInterface::class, $class);
+        }
+
+        return $this->getContainer()->get(HandlerInterface::class);
     }
 
     /**
@@ -189,6 +258,72 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     }
 
     /**
+     * Return the router.
+     *
+     * @return \League\Route\RouteCollection
+     */
+    public function getRouter()
+    {
+        if (!isset($this->router)) {
+            $this->router = (new RouteCollection($this->getContainer()));
+        }
+
+        return $this->router;
+    }
+
+    /**
+     * Get the request
+     *
+     * @return ServerRequestInterface
+     */
+    public function getRequest()
+    {
+        if (!$this->getContainer()->has(ServerRequestInterface::class)) {
+            $this->getContainer()->share(ServerRequestInterface::class, ServerRequestFactory::fromGlobals());
+        }
+
+        return $this->getContainer()->get(ServerRequestInterface::class);
+    }
+
+    /**
+     * Get the response
+     *
+     * @param string $content
+     * @return ResponseInterface
+     */
+    public function getResponse($content = '')
+    {
+        //transform content by environment and request type
+        if ($this->isAjax()) {
+            if ($content instanceof JsonResponse) {
+                $content = json_decode($content->getBody());
+            } elseif (!is_array($content)) {
+                $content = [$content];
+            }
+        } else {
+            if (is_array($content)) {
+                $content = implode('', $content);
+            } elseif ($content instanceof ResponseInterface) {
+                $content = $content->getBody()->__toString();
+            }
+        }
+        if (!$this->getContainer()->has(ResponseInterface::class)) {
+            if ($this->isCli()) {
+                $class = TextResponse::class;
+            } elseif ($this->isAjax()) {
+                $class = JsonResponse::class;
+
+            } else {
+                $class = HtmlResponse::class;
+            }
+            $this->getContainer()->add(ResponseInterface::class, $class);
+        }
+
+        return $this->getContainer()->get(ResponseInterface::class, [$content]);
+    }
+
+
+    /**
      * Get response emitter
      *
      * @return \Zend\Diactoros\Response\EmitterInterface
@@ -203,30 +338,37 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     }
 
     /**
-     * Set response emitter
+     * Check if request is a ajax request
      *
-     * @param \Zend\Diactoros\Response\EmitterInterface $responseEmitter
-     * @return $this
+     * @return bool
      */
-    public function setResponseEmitter($responseEmitter)
+    public function isAjax()
     {
-        $this->responseEmitter = $responseEmitter;
-
-        return $this;
+        return
+            false !== strpos(
+                strtolower(ServerRequestFactory::getHeader('x-requested-with', $this->getRequest()->getHeaders(), '')),
+                'xmlhttprequest'
+            ) && $this->isHttp();
     }
 
     /**
-     * Set the exception decorator.
+     * Check server environment for cli
      *
-     * @param callable $callback
-     *
-     * @return $this
+     * @return bool
      */
-    public function setExceptionDecorator(callable $callback)
+    public function isCli()
     {
-        $this->exceptionDecorator = $callback;
+        return php_sapi_name() === 'cli';
+    }
 
-        return $this;
+    /**
+     * Check server environment for http
+     *
+     * @return bool
+     */
+    public function isHttp()
+    {
+        return !$this->isCli();
     }
 
     /**
@@ -320,21 +462,74 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     {
 
         // Passes the request to the container
-        $this->getContainer()->add(ServerRequestInterface::class, $request);
+        $this->getContainer()->share(ServerRequestInterface::class, $request);
 
         try {
-
-            $response = $this->handleRequest($request);
-
-        } catch (\Exception $e) {
-
-            if (!$catch) {
-                throw $e;
-            }
-
-            $response = $this->handleError($request, $e);
+            $response = $this->handleRequest($request, $this->getResponse());
+        } catch (\Exception $exception) {
+            $response = $this->handleError($request, $exception, $catch);
+            $response = $response->withStatus($exception instanceof NotFoundException ? 404 : 500);
 
         }
+
+        return $response;
+    }
+
+    /**
+     * @param ServerRequestInterface $request
+     * @param ResponseInterface $response
+     * @return ResponseInterface
+     */
+    public function handleRequest(ServerRequestInterface $request, ResponseInterface $response)
+    {
+        $this->emit('request.received', $request);
+
+        $response = $this->getRouter()->dispatch(
+            $request,
+            $response
+        );
+
+        $this->emit('response.created', $request, $response);
+
+        return $response;
+    }
+
+    /**
+     * @param ServerRequestInterface $request
+     * @param \Throwable|\Exception $exception
+     * @param bool $catch
+     *
+     * @return ResponseInterface
+     *
+     * @throws \Throwable
+     */
+    public function handleError(ServerRequestInterface $request, $exception, $catch = true)
+    {
+        $errorHandler = new Run();
+        $errorHandler->pushHandler($this->getErrorResponseHandler());
+        if (false === $this->getConfig('error', static::CONFIG_ERROR)) {
+            $message = $exception->getMessage();
+        } else {
+            $errorHandler->allowQuit($this->error);
+
+            if (!$catch) {
+                $errorHandler->register();
+                throw $exception;
+            }
+
+            $method = Run::EXCEPTION_HANDLER;
+            ob_start();
+            $errorHandler->$method($exception);
+            $message = ob_get_clean();
+        }
+        $response = $this->getResponse();
+        $this->emit('response.error', $request, $response, $exception);
+
+        if (empty($response->getBody()->__toString())) {
+            $response->getBody()->write($message);
+        }
+
+        $this->error = true;
 
         return $response;
     }
@@ -442,63 +637,5 @@ class Application implements ApplicationInterface, ContainerAwareInterface, Http
     public function register($serviceProvider)
     {
         $this->getContainer()->addServiceProvider($serviceProvider);
-    }
-
-    /**
-     * Set a config item
-     *
-     * @param string $key
-     * @param mixed $value
-     */
-    public function setConfig($key, $value)
-    {
-        $this->config[$key] = $value;
-    }
-
-    /**
-     * Get a config key's value
-     *
-     * @param string $key
-     * @param mixed $default
-     *
-     * @return mixed
-     */
-    public function getConfig($key, $default = null)
-    {
-        return isset($this->config[$key]) ? $this->config[$key] : $default;
-    }
-
-    /**
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     */
-    public function handleRequest(ServerRequestInterface $request)
-    {
-        $this->emit('request.received', $request);
-
-        $response = $this->getRouter()->dispatch(
-            $request,
-            new HtmlResponse('')
-        );
-
-        $this->emit('response.created', $request, $response);
-        return $response;
-    }
-
-    /**
-     * @param ServerRequestInterface $request
-     * @param $e
-     * @return mixed
-     * @throws
-     */
-    public function handleError(ServerRequestInterface $request, $e)
-    {
-        $response = call_user_func($this->exceptionDecorator, $e);
-        if (!$response instanceof ResponseInterface) {
-            throw new \LogicException('Exception decorator did not return an instance of ' . ResponseInterface::class);
-        }
-
-        $this->emit('response.created', $request, $response);
-        return $response;
     }
 }